Refactor auto-join configuration UI
- Move MultiDomainAutoJoinModal to dedicated file
- Add batch domain update API endpoint (domainUpdates array)
- Use arrow function for handleSave
- Conditionally mount only the relevant modal based on isMultiDomain
