+++ front/components/workspace/sso/AutoJoinToggle.tsx
@@ -1,18 +1,17 @@
 import {
   Button,
-  Checkbox,
   Dialog,
   DialogContainer,
   DialogContent,
   DialogFooter,
   DialogHeader,
   DialogTitle,
-  Label,
   Page,
 } from "@dust-tt/sparkle";
 import type { Organization } from "@workos-inc/node";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 
+import { MultiDomainAutoJoinModal } from "@app/components/workspace/sso/MultiDomainAutoJoinModal";
 import { UpgradePlanDialog } from "@app/components/workspace/UpgradePlanDialog";
 import { useSendNotification } from "@app/hooks/useNotification";
 import { clientFetch } from "@app/lib/egress/client";
@@ -110,160 +109,6 @@ function DomainAutoJoinModal({
   );
 }
 
-interface MultiDomainAutoJoinModalProps {
-  workspaceVerifiedDomains: WorkspaceDomain[];
-  isOpen: boolean;
-  onClose: () => void;
-  owner: WorkspaceType;
-}
-
-function MultiDomainAutoJoinModal({
-  workspaceVerifiedDomains,
-  isOpen,
-  onClose,
-  owner,
-}: MultiDomainAutoJoinModalProps) {
-  const sendNotification = useSendNotification();
-  const [domainOverrides, setDomainOverrides] = useState<
-    Record<string, boolean>
-  >({});
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
-  useEffect(() => {
-    if (isOpen) {
-      setDomainOverrides({});
-    }
-  }, [isOpen]);
-
-  const hasChanges = workspaceVerifiedDomains.some((d) => {
-    const desiredValue = domainOverrides[d.domain];
-    return (
-      desiredValue !== undefined && desiredValue !== d.domainAutoJoinEnabled
-    );
-  });
-
-  const isDomainEnabled = (domain: WorkspaceDomain): boolean =>
-    domainOverrides[domain.domain] ?? domain.domainAutoJoinEnabled;
-
-  const handleDomainToggle = (domain: WorkspaceDomain) => {
-    setDomainOverrides((prev) => {
-      const currentValue = prev[domain.domain] ?? domain.domainAutoJoinEnabled;
-      return {
-        ...prev,
-        [domain.domain]: !currentValue,
-      };
-    });
-  };
-
-  async function handleSave(): Promise<void> {
-    const updates = workspaceVerifiedDomains
-      .map((d) => ({
-        domain: d.domain,
-        enabled: domainOverrides[d.domain] ?? d.domainAutoJoinEnabled,
-        previous: d.domainAutoJoinEnabled,
-      }))
-      .filter((u) => u.enabled !== u.previous)
-      .map(({ domain, enabled }) => ({ domain, enabled }));
-
-    if (updates.length === 0) {
-      onClose();
-      return;
-    }
-
-    setIsSubmitting(true);
-    try {
-      const results = await Promise.allSettled(
-        updates.map((u) =>
-          clientFetch(`/api/w/${owner.sId}`, {
-            method: "POST",
-            headers: {
-              "Content-Type": "application/json",
-            },
-            body: JSON.stringify({
-              domain: u.domain,
-              domainAutoJoinEnabled: u.enabled,
-            }),
-          })
-        )
-      );
-
-      const failedDomains: string[] = [];
-      results.forEach((result, index) => {
-        if (result.status === "rejected" || !result.value.ok) {
-          failedDomains.push(updates[index].domain);
-        }
-      });
-
-      if (failedDomains.length > 0) {
-        sendNotification({
-          type: "error",
-          title: "Update failed",
-          description: `Failed to update auto-join for: ${failedDomains.map((d) => `@${d}`).join(", ")}`,
-        });
-      }
-
-      // Full refresh to update owner object and keep formValidation logic working.
-      window.location.reload();
-    } finally {
-      setIsSubmitting(false);
-      onClose();
-    }
-  }
-
-  return (
-    <Dialog
-      open={isOpen}
-      onOpenChange={(open) => {
-        if (!open && !isSubmitting) {
-          onClose();
-        }
-      }}
-    >
-      <DialogContent size="md">
-        <DialogHeader>
-          <DialogTitle>Configure Auto-join</DialogTitle>
-        </DialogHeader>
-        <DialogContainer>
-          <div className="flex flex-col gap-4">
-            <p className="text-sm text-muted-foreground">
-              Select which domains should allow users to automatically join your
-              workspace when they sign up with a matching email address.
-            </p>
-            <div className="flex flex-col gap-3">
-              {workspaceVerifiedDomains.map((d) => (
-                <div key={d.domain} className="flex items-center gap-2">
-                  <Checkbox
-                    id={`domain-${d.domain}`}
-                    checked={isDomainEnabled(d)}
-                    onCheckedChange={() => handleDomainToggle(d)}
-                    disabled={isSubmitting}
-                  />
-                  <Label htmlFor={`domain-${d.domain}`} className="font-normal">
-                    @{d.domain}
-                  </Label>
-                </div>
-              ))}
-            </div>
-          </div>
-        </DialogContainer>
-        <DialogFooter
-          leftButtonProps={{
-            label: "Cancel",
-            variant: "outline",
-            disabled: isSubmitting,
-          }}
-          rightButtonProps={{
-            label: isSubmitting ? "Saving..." : "Save",
-            variant: "primary",
-            onClick: handleSave,
-            disabled: !hasChanges || isSubmitting,
-          }}
-        />
-      </DialogContent>
-    </Dialog>
-  );
-}
-
 type AutoJoinToggleProps = {
   domains: Organization["domains"];
   workspaceVerifiedDomains: WorkspaceDomain[];
@@ -291,22 +136,23 @@ export function AutoJoinToggle({
 
   return (
     <>
-      <DomainAutoJoinModal
-        domainAutoJoinEnabled={domainAutoJoinEnabled}
-        isOpen={isActivateAutoJoinOpened && !isMultiDomain}
-        onClose={() => {
-          setIsActivateAutoJoinOpened(false);
-        }}
-        domains={domains}
-        owner={owner}
-      />
-      {isMultiDomain && (
+      {isMultiDomain ? (
         <MultiDomainAutoJoinModal
           workspaceVerifiedDomains={workspaceVerifiedDomains}
           isOpen={isActivateAutoJoinOpened}
           onClose={() => {
             setIsActivateAutoJoinOpened(false);
           }}
+          owner={owner}
+        />
+      ) : (
+        <DomainAutoJoinModal
+          domainAutoJoinEnabled={domainAutoJoinEnabled}
+          isOpen={isActivateAutoJoinOpened}
+          onClose={() => {
+            setIsActivateAutoJoinOpened(false);
+          }}
+          domains={domains}
           owner={owner}
         />
       )}
